index.php
Saving a file uploaded by the user is one of the frequent operations that one needs to do when designing user interfaces. There are several ways to do this, but it is often preferred to save the files directly within a specific folder and in the database indicate the path where to retrieve it.